Set up routers and controllers properly
The idea is that everything pertaining to
users will be accessed through usersRouter.js
which will be placed in app.js, for example:

app.use("/", usersRouter);

Then usersRouter.js will be placed in
the routes folder and all user routers
placed in the router js file like this:

app.get("/", usersControllers.usersListGet);
app.get("/users", usersControllers.usersFormGet);

Note that a usersController file will be created
and placed in the controllers folder. All user
logic will be placed in that file, for example:

Render index (Homepage):
exports.usersListGet = (req, res, next) => {
    res.render("index", { title: "Users List" });
}

Render user form:
exports.usersFormGet = (req, res, next) => {
    res.render("userForm", { title: "User Form" });
}

Thus the organization is as such that the routers
direct the requests to the controllers which
contain the logic to handle the requests. More
explicitly, methods containing the required logic
are placed in the controllers which are javascript
files and these methods are accessed using .notation,
for example: usersControllers.usersListGet.